Solve circular dep
