+++ ee/vellum_ee/workflows/display/utils/vellum.py
@@ -1,31 +1,8 @@
-import enum
-import json
-import typing
-from typing import Any, List, Union, cast
+from typing import Any, TypeVar
 
-from vellum import ChatMessage, SearchResult, SearchResultRequest, VellumVariableType
-from vellum.client.types.array_vellum_value import ArrayVellumValue
-from vellum.client.types.array_vellum_value_request import ArrayVellumValueRequest
-from vellum.client.types.audio_vellum_value import AudioVellumValue
-from vellum.client.types.audio_vellum_value_request import AudioVellumValueRequest
-from vellum.client.types.chat_history_vellum_value import ChatHistoryVellumValue
-from vellum.client.types.chat_history_vellum_value_request import ChatHistoryVellumValueRequest
-from vellum.client.types.error_vellum_value import ErrorVellumValue
-from vellum.client.types.error_vellum_value_request import ErrorVellumValueRequest
-from vellum.client.types.function_call_vellum_value import FunctionCallVellumValue
-from vellum.client.types.function_call_vellum_value_request import FunctionCallVellumValueRequest
-from vellum.client.types.image_vellum_value import ImageVellumValue
-from vellum.client.types.image_vellum_value_request import ImageVellumValueRequest
-from vellum.client.types.json_vellum_value import JsonVellumValue
-from vellum.client.types.json_vellum_value_request import JsonVellumValueRequest
-from vellum.client.types.number_vellum_value import NumberVellumValue
-from vellum.client.types.number_vellum_value_request import NumberVellumValueRequest
-from vellum.client.types.search_results_vellum_value import SearchResultsVellumValue
-from vellum.client.types.search_results_vellum_value_request import SearchResultsVellumValueRequest
-from vellum.client.types.string_vellum_value import StringVellumValue
-from vellum.client.types.string_vellum_value_request import StringVellumValueRequest
-from vellum.client.types.vellum_value import VellumValue
+from vellum.client.types.vellum_variable_type import VellumVariableType
 from vellum.workflows.descriptors.base import BaseDescriptor
+from vellum.workflows.nodes.displayable.bases.utils import primitive_to_vellum_value
 from vellum.workflows.references import OutputReference, WorkflowInputReference
 from vellum.workflows.references.execution_count import ExecutionCountReference
 from vellum.workflows.references.node import NodeReference
@@ -46,7 +23,7 @@ from vellum_ee.workflows.display.vellum import (
     WorkspaceSecretPointer,
 )
 
-_T = typing.TypeVar("_T")
+_T = TypeVar("_T")
 
 
 def infer_vellum_variable_type(value: Any) -> VellumVariableType:
@@ -105,67 +82,3 @@ def create_node_input_value_pointer_rule(
         return ConstantValuePointer(type="CONSTANT_VALUE", data=vellum_value)
 
     raise ValueError(f"Unsupported descriptor type: {value.__class__.__name__}")
-
-
-def primitive_to_vellum_value(value: Any) -> VellumValue:
-    """Converts a python primitive to a VellumValue"""
-
-    if isinstance(value, str):
-        return StringVellumValue(value=value)
-    elif isinstance(value, enum.Enum):
-        return StringVellumValue(value=value.value)
-    elif isinstance(value, (int, float)):
-        return NumberVellumValue(value=value)
-    elif isinstance(value, list) and (
-        all(isinstance(message, ChatMessage) for message in value)
-        or all(isinstance(message, ChatMessage) for message in value)
-    ):
-        chat_messages = cast(Union[List[ChatMessage], List[ChatMessage]], value)
-        return ChatHistoryVellumValue(value=chat_messages)
-    elif isinstance(value, list) and (
-        all(isinstance(search_result, SearchResultRequest) for search_result in value)
-        or all(isinstance(search_result, SearchResult) for search_result in value)
-    ):
-        search_results = cast(Union[List[SearchResultRequest], List[SearchResult]], value)
-        return SearchResultsVellumValue(value=search_results)
-    elif isinstance(
-        value,
-        (
-            StringVellumValue,
-            NumberVellumValue,
-            JsonVellumValue,
-            ImageVellumValue,
-            AudioVellumValue,
-            FunctionCallVellumValue,
-            ErrorVellumValue,
-            ArrayVellumValue,
-            ChatHistoryVellumValue,
-            SearchResultsVellumValue,
-        ),
-    ):
-        return value
-    elif isinstance(
-        value,
-        (
-            StringVellumValueRequest,
-            NumberVellumValueRequest,
-            JsonVellumValueRequest,
-            ImageVellumValueRequest,
-            AudioVellumValueRequest,
-            FunctionCallVellumValueRequest,
-            ErrorVellumValueRequest,
-            ArrayVellumValueRequest,
-            ChatHistoryVellumValueRequest,
-            SearchResultsVellumValueRequest,
-        ),
-    ):
-        # This type ignore is safe because consumers of this function won't care the difference between
-        # XVellumValue and XVellumValueRequest. Hopefully in the near future, neither will we
-        return value  # type: ignore
-
-    try:
-        json_value = json.dumps(value)
-    except json.JSONDecodeError:
-        raise ValueError(f"Unsupported variable type: {value.__class__.__name__}")
-
-    return JsonVellumValue(value=json.loads(json_value))
